+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.22] - 2024-07-23
+
+### Added
+
+-   Implemented a pause feature that allows players to pause and resume the game.
+-   Moved the music toggle and volume control to the pause menu for better accessibility.
+
 ## [1.21] - 2024-07-23
 
 ### Changed